+++ navigation/sample-standard/src/main/java/ru/surfstudio/android/navigation/sample_standard/screen/bottom_nav/profile/ProfileFragmentView.kt
@@ -23,5 +23,6 @@ class ProfileFragmentView : BaseRxFragmentView() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?, viewRecreated: Boolean) {
         profile_settings_btn.setOnClickListener { bm.openSettings.accept() }
         profile_logout_btn.setOnClickListener { bm.openConfirmLogoutScreen.accept() }
+        profile_search_btn.setOnClickListener { bm.openSearch.accept() }
     }
 }