+++ navigation/sample-standard/src/main/java/ru/surfstudio/android/navigation/sample_standard/screen/bottom_nav/profile/ProfilePresenter.kt
@@ -6,6 +6,7 @@ import ru.surfstudio.android.dagger.scope.PerScreen
 import ru.surfstudio.android.navigation.command.activity.Start
 import ru.surfstudio.android.navigation.command.dialog.Show
 import ru.surfstudio.android.navigation.command.fragment.RemoveLast
+import ru.surfstudio.android.navigation.command.fragment.Replace
 import ru.surfstudio.android.navigation.command.fragment.base.FragmentNavigationCommand.Companion.ACTIVITY_NAVIGATION_TAG
 import ru.surfstudio.android.navigation.executor.NavigationCommandExecutor
 import ru.surfstudio.android.navigation.observer.ScreenResultObserver
@@ -14,6 +15,9 @@ import ru.surfstudio.android.navigation.sample_standard.screen.base.dialog.simpl
 import ru.surfstudio.android.navigation.sample_standard.screen.base.presenter.CommandExecutionPresenter
 import ru.surfstudio.android.navigation.sample_standard.screen.bottom_nav.profile.logout.LogoutConfirmationRoute
 import ru.surfstudio.android.navigation.sample_standard.screen.bottom_nav.profile.settings.ApplicationSettingsRoute
+import ru.surfstudio.android.navigation.sample_standard.screen.search.request.SearchRequestRoute
+import ru.surfstudio.android.navigation.sample_standard.screen.search.results.SearchResultRoute
+import ru.surfstudio.android.navigation.sample_standard.utils.animations.ModalAnimations
 import javax.inject.Inject
 
 @PerScreen
@@ -27,8 +31,22 @@ class ProfilePresenter @Inject constructor(
     override fun onFirstLoad() {
         val confirmLogoutRoute = LogoutConfirmationRoute()
         subscribeToLogoutResult(confirmLogoutRoute)
+        subscribeOnSearchResults()
         bm.openConfirmLogoutScreen bindTo { Show(confirmLogoutRoute).execute() }
         bm.openSettings.bindTo { Start(ApplicationSettingsRoute()).execute() }
+        bm.openSearch.bindTo {
+            Replace(
+                    route = SearchRequestRoute(),
+                    animations = ModalAnimations,
+                    sourceTag = ACTIVITY_NAVIGATION_TAG
+            ).execute()
+        }
+    }
+
+    private fun subscribeOnSearchResults() {
+        subscribe(screenResultObserver.observeScreenResult(SearchRequestRoute())){ result: String ->
+            Replace(SearchResultRoute(result)).execute()
+        }
     }
 
     private fun subscribeToLogoutResult(confirmLogoutRoute: LogoutConfirmationRoute) {