+++ navigation/sample/src/main/java/ru/surfstudio/android/navigation/sample/app/screen/main/home/HomeTabFragment.kt
@@ -10,7 +10,6 @@ import ru.surfstudio.android.navigation.command.fragment.Replace
 import ru.surfstudio.android.navigation.sample.R
 import ru.surfstudio.android.navigation.sample.app.App
 import ru.surfstudio.android.navigation.sample.app.screen.main.home.nested.HomeNestedRoute
-import ru.surfstudio.android.navigation.sample.app.utils.animations.FadeAnimations
 
 class HomeTabFragment : Fragment() {
 
@@ -20,6 +19,6 @@ class HomeTabFragment : Fragment() {
 
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
-        home_add_btn.setOnClickListener { App.navigator.execute(Replace(HomeNestedRoute(), sourceTag = tag!!)) }
+        home_add_btn.setOnClickListener { App.navCommandExecutor.execute(Replace(HomeNestedRoute(), sourceTag = tag!!)) }
     }
 }