+++ navigation/sample/src/main/java/ru/surfstudio/android/navigation/sample/app/screen/main/home/nested/HomeNestedFragment.kt
@@ -21,7 +21,7 @@ class HomeNestedFragment : Fragment() {
         val currentOrder = HomeNestedRoute(arguments).order
         home_nested_tv.text = "Home\nNested\nFragment\n#$currentOrder"
         val nextOrder = currentOrder + 1
-        home_nested_add_btn.setOnClickListener { App.navigator.execute(Replace(HomeNestedRoute(nextOrder), sourceTag = tag!!)) }
+        home_nested_add_btn.setOnClickListener { App.navCommandExecutor.execute(Replace(HomeNestedRoute(nextOrder), sourceTag = tag!!)) }
     }
 
 }