+++ navigation/sample/src/main/java/ru/surfstudio/android/navigation/sample/app/screen/splash/SplashFragment.kt
@@ -10,6 +10,7 @@ import ru.surfstudio.android.navigation.command.fragment.ReplaceHard
 import ru.surfstudio.android.navigation.sample.R
 import ru.surfstudio.android.navigation.sample.app.App
 import ru.surfstudio.android.navigation.sample.app.screen.auth.AuthRoute
+import ru.surfstudio.android.navigation.sample.app.utils.animations.FadeAnimations
 
 class SplashFragment : Fragment() {
 
@@ -20,7 +21,7 @@ class SplashFragment : Fragment() {
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
         Handler().postDelayed({
-            App.navigator.execute(ReplaceHard(AuthRoute()))
+            App.navCommandExecutor.execute(ReplaceHard(AuthRoute(), animations = FadeAnimations))
         }, 1500L)
     }
 }