+++ navigation/sample/src/main/java/ru/surfstudio/android/navigation/sample/app/utils/animations/ModalAnimations.kt
@@ -3,13 +3,9 @@ package ru.surfstudio.android.navigation.sample.app.utils.animations
 import ru.surfstudio.android.navigation.animation.resource.BaseResourceAnimations
 import ru.surfstudio.android.navigation.sample.R
 
-
-/**
- * Набор анимаций Fade In/Out для переключения фрагментов.
- */
-object FadeAnimations : BaseResourceAnimations(
-        enterAnimation = R.anim.fade_in_fast,
+object ModalAnimations: BaseResourceAnimations(
+        enterAnimation = R.anim.slide_in_from_bottom,
         exitAnimation = R.anim.fade_out_fast,
         popEnterAnimation = R.anim.fade_in_fast,
-        popExitAnimation = R.anim.fade_out_fast
+        popExitAnimation = R.anim.slide_out_to_bottom
 )