+++ navigation/RELEASE_NOTES.md
@@ -13,4 +13,5 @@
 * ANDDEP-1154 Fixed crashes "You must specify unique tag"
 * ANDDEP-1153 Fixed wrong behavior RemoveLast
 * ANDDEP-1100 Added command StartForResult to open system activities for getting result
+* ANDDEP-1103 Added sequential async and sync command execution
 * ANDDEP-1148 code updated for target sdk 30